main.py decomposed into project folder
